Move getPosition hook out of Navbar
